refactor(ui): migrate last legacy alert and select consumers to shadcn
